+++ src/build.task.ts
@@ -6,7 +6,7 @@ import type {GroConfig} from './config/config.js';
 import type {TaskEvents as ServerTaskEvents} from './server.task.js';
 import {printTimings} from './utils/print.js';
 import {toArray} from './utils/array.js';
-import type {AdaptBuildsContext} from './config/adapt.js';
+import type {AdaptBuildsContext} from './adapt/adapter.js';
 import {buildSourceDirectory} from './build/buildSourceDirectory.js';
 
 // outputs build artifacts to dist/ using SvelteKit or Gro config
